third_party/WebKit/LayoutTests/plugins/request-low-latency-touch.html

<!DOCTYPE html>
<html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<body>
<embed id="touch_plugin" type="application/x-webkit-test-webplugin" accepts-touch="raw-lowlatency"></embed>
<script>

test(function() {
    // ensure layout is valid.
    touch_plugin.offsetLeft;

    assert_equals(internals.touchStartOrMoveEventHandlerCount(document), 1, "One touch event handler registered.");
}, "raw handler added");

</script>
</body>
